refactor(types): authFormField enum and type
