--- src/types/auth-form-field-enum.ts
@@ -1,7 +1,0 @@
-enum AuthFormFieldEnum {
-  USERNAME,
-  PASSWORD,
-  VERIFICATION_CODE,
-}
-
-export default AuthFormFieldEnum;
