+++ src/types/authFormField.ts
@@ -1,11 +1,13 @@
-import AuthFormFieldEnum from './auth-form-field-enum';
+export enum AuthFormFieldEnum {
+  USERNAME,
+  PASSWORD,
+  VERIFICATION_CODE,
+}
 
-type AuthFormFieldType = {
+export type AuthFormFieldType = {
   label: string;
   type?: AuthFormFieldEnum;
   fixedValue?: boolean;
   currentValue: string;
   onChange?: (input: string) => void;
 };
-
-export default AuthFormFieldType;
